Added Authentication System
On the backend I used Entity Framework Core and Identity Core.  The DB is set to InMemory.  Though can easily be configured to connect to a SQL DB.

ClientApp has a home page with links to login or register.  Users can signup and then login, which grants them access to the Dashboard.  The Dashboard has a authGuard preventing unregistered users from accessing it.

I updated the Angular app to Angular 7 and installed Bootstrap 4.